Rest first commit

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -4,6 +4,8 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import ru.kata.spring.boot_security.demo.dao.UserDao;
+import ru.kata.spring.boot_security.demo.exceptions.UserAlreadyExistException;
+import ru.kata.spring.boot_security.demo.exceptions.UserNotExistException;
 import ru.kata.spring.boot_security.demo.model.User;
 
 import javax.transaction.Transactional;
@@ -22,8 +24,13 @@ public class UserServiceImpl implements UserService{
     @Transactional
     @Override
     public void add(User user) {
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        userDao.add(user);
+        if (getByUsername(user.getUsername()) == null) {
+            user.setPassword(passwordEncoder.encode(user.getPassword()));
+            userDao.add(user);
+        } else {
+            throw new UserAlreadyExistException("User already exist");
+        }
+
     }
 
     @Override
@@ -49,14 +56,22 @@ public class UserServiceImpl implements UserService{
     @Transactional
     @Override
     public void update(User user) {
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        userDao.update(user);
+        if (getById(user.getId()) != null) {
+            user.setPassword(passwordEncoder.encode(user.getPassword()));
+            userDao.update(user);
+        } else {
+            throw new UserNotExistException("User not found");
+        }
     }
 
     @Transactional
     @Override
     public void delete(User user) {
-        userDao.delete(user);
+        if (getById(user.getId()) != null) {
+            userDao.delete(user);
+        } else {
+            throw new UserNotExistException("User not found");
+        }
     }
 
 }
